requirements.txt
altair @ file:///tmp/build/80754af9/altair_1599835197802/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blinker==1.4
blis==0.7.8
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
catalogue==1.0.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chardet==5.0.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056591685/work
commonmark @ file:///Users/ktietz/demo/mc3/conda-bld/commonmark_1630649545323/work
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cymem @ file:///opt/conda/conda-bld/cymem_1651237256138/work
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
future==0.18.2
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///tmp/build/80754af9/gitpython_1624609474443/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646983213839/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema==3.0.2
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
jupyter_client @ file:///opt/conda/conda-bld/jupyter_client_1661848916004/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
langcodes==3.3.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
murmurhash==1.0.8
nbclient @ file:///tmp/build/80754af9/nbclient_1650308370114/work
nbconvert @ file:///opt/conda/conda-bld/nbconvert_1649751911790/work
nbformat @ file:///tmp/build/80754af9/nbformat_1649844918397/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847911453/work
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
notebook @ file:///tmp/abs_abf6xa6h6f/croots/recipe/notebook_1659083654985/work
numexpr @ file:///opt/conda/conda-bld/numexpr_1656940300424/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathy==0.6.2
pdfminer.six==20220524
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
plac @ file:///tmp/build/80754af9/plac_1594259967326/work
preshed==3.0.7
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==8.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.15.0
pydantic==1.9.2
pydeck @ file:///tmp/abs_c06301fd-7800-4748-8e9e-0f9ab34bc0cess8ywq6h/croots/recipe/pydeck_1657894564709/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
Pympler @ file:///tmp/build/80754af9/pympler_1602785470644/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyresparser==1.0.6
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
regex @ file:///tmp/abs_41f5bce5-0a2e-45aa-b231-1fd2fbd57753gfpe6sjm/croots/recipe/regex_1658257178822/work
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rich @ file:///tmp/abs_f016wesp06/croots/recipe/rich_1659708287203/work
semver @ file:///tmp/build/80754af9/semver_1603822362442/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sortedcontainers==2.4.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
spacy @ file:///tmp/build/80754af9/spacy_1608321093591/work
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly @ file:///tmp/build/80754af9/srsly_1607548514031/work
streamlit @ file:///tmp/abs_c4af64a6-e8cd-486a-8e7a-4fa8b9d0c903olpgeyyl/croots/recipe/streamlit_1657914742590/work
terminado @ file:///tmp/build/80754af9/terminado_1644304459949/work
testpath @ file:///opt/conda/conda-bld/testpath_1655908557405/work
thinc @ file:///tmp/build/80754af9/thinc_1607710140838/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1650891076910/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typer==0.4.2
typing_extensions==4.1.1
tzlocal @ file:///tmp/build/80754af9/tzlocal_1597435723902/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
validators @ file:///tmp/build/80754af9/validators_1612286467315/work
wasabi==0.10.1
watchdog @ file:///tmp/build/80754af9/watchdog_1638367288901/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354203/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
